requirements.txt
alembic revision --autogenerate -m "initial_schema"# Core framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# GraphQL
strawberry-graphql[fastapi]==0.243.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Configuration
# Pin pydantic <2.11 — strawberry 0.243.1 references pydantic internals
# (`is_new_type`) that were removed in pydantic 2.11.
pydantic>=2.6,<2.11
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Utilities
email-validator==2.2.0
